refactor: extract PIR helpers to spgrep.symmetry.pir
Move the physically-irreducible-representation code out of
spgrep.symmetry.enumerate into a sibling module. No behavior change.

- New spgrep/symmetry/pir.py with purify_real_irrep_value and a
  private helper _make_physically_irreps_from_complex that absorbs
  the near-duplicate real=True blocks from enumerate_small_repre-
  sentations and enumerate_unitary_irreps.
- enumerate_small_representations and enumerate_unitary_irreps now
  delegate their real=True branch to the helper, passing kpoint or
  None to toggle the commensurate-2k indicator override.
- Drop the PIR-only imports from enumerate.py.
- Repoint the deprecated spgrep.irreps shim at the new module for
  purify_real_irrep_value.

Motivates the follow-up PIR-on-G^{k,k-bar} work that would otherwise
add another ~250 lines of PIR-specific helpers to enumerate.py.
